The basic test file is moved to top directory because of need of testing it.
